requirements.txt
httpx>=0.25.0
python-telegram-bot>=20.0
rugcheck>=1.0.0
python-dotenv>=1.0.0
